% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_to_dataframe.R
\name{array_to_df}
\alias{array_to_df}
\title{Long format from Array}
\usage{
array_to_df(a, r)
}
\arguments{
\item{a}{Array of the form p x p x m}

\item{r}{optional scale vector used in computation}
}
\value{
data.frame
}
\description{
Long format from Array
}
\details{
The functions in this package return arrays of dimensions p x p x m, where p = number of types, m = number of scales.
This function turns this such arrays into data frames with columns (i, j, r, v). If `r` is provided,
the `r` in the output should match. Otherwise, output `r` is the index.
}
