% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pp.R
\name{check_mpp}
\alias{check_mpp}
\title{check pattern format}
\usage{
check_mpp(x, marked = TRUE)
}
\arguments{
\item{x}{pattern candidate}

\item{marked}{whether to parse the marks as well?}
}
\description{
check pattern format
}

