% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iD-cross-all.R
\name{iD_cross_all_box}
\alias{iD_cross_all_box}
\title{quick inhomogeneous cross D, minus border correction}
\usage{
iD_cross_all_box(x, r, int, correction = TRUE)
}
\arguments{
\item{x}{point pattern, multivariate}

\item{r}{range vector}

\item{int}{vector of intensities per point}

\item{correction}{use or not use reduced sample border correction}
}
\description{
Nearest neighbour distance distribution function, assumes rectangnular window.
}
