% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iD-cross-all.R
\name{iD_cross_all_box_from}
\alias{iD_cross_all_box_from}
\title{quick inhomogeneous cross D, minus border correction, one-to-many}
\usage{
iD_cross_all_box_from(x, r, from, int, correction = TRUE)
}
\arguments{
\item{x}{point pattern, multivariate}

\item{r}{range vector}

\item{from}{target type, integer in mark-levels}

\item{int}{vector of intensities per point}

\item{correction}{use or not use reduced sample border correction}
}
\description{
Nearest neighbour distance distribution function, assumes rectangnular window.
}
