% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iISAR.R
\name{iISAR_all_box}
\alias{iISAR_all_box}
\title{Quick ISAR function, all}
\usage{
iISAR_all_box(x, r, int, correction = TRUE)
}
\arguments{
\item{x}{point pattern, multivariate}

\item{r}{range vector}

\item{int}{vector of intensities per point}

\item{correction}{use or not use reduced sample border correction}
}
\description{
ISAR, inhomog.
}
