% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iISAR.R
\name{iISAR_all_box_from}
\alias{iISAR_all_box_from}
\title{Quick ISAR function, target type}
\usage{
iISAR_all_box_from(x, r, from, int, correction = TRUE)
}
\arguments{
\item{x}{point pattern, multivariate}

\item{r}{range vector}

\item{from}{target type, integer}

\item{int}{vector of intensities per point}

\item{correction}{use or not use reduced sample border correction}
}
\description{
ISAR, inhomog.
}
