% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iK-cross-all.R
\name{iK_box}
\alias{iK_box}
\title{quick K, border correction
assumes rectangnular window}
\usage{
iK_box(x, r, int, correction = TRUE)
}
\arguments{
\item{x}{point pattern}

\item{r}{range vector}

\item{int}{vector of intensities per point}

\item{correction}{use or not use translation correction}
}
\description{
quick K, border correction
assumes rectangnular window
}

