% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iK-cross-all.R
\name{iK_cross_all_box}
\alias{iK_cross_all_box}
\title{quick cross K, translation border correction
assumes rectangnular window}
\usage{
iK_cross_all_box(x, r, int, correction = TRUE)
}
\arguments{
\item{x}{point pattern, multivariate}

\item{r}{range vector}

\item{int}{vector of intensities per point}

\item{correction}{use or not use translation correction}
}
\description{
quick cross K, translation border correction
assumes rectangnular window
}

