% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity-estimators.R
\name{intensity_adapted}
\alias{intensity_adapted}
\title{Intensity estimator for 2D pattern in rectangle}
\usage{
intensity_adapted(x, r, type = "all")
}
\arguments{
\item{x}{point pattern with rectangluar window}

\item{r}{range parameter in case needed}

\item{type}{estimator to use.}
}
\description{
Adapted homogeneous intensity estimators for use with other estimators.
}
\details{
type:

* "n": none
* "b": minus sampling
* "v": relative area
* "s": relative border length

as in IPSS'08 p 193-194.
}
