% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox-disc-intersections.R
\name{intersect_bbox_with_disc}
\alias{intersect_bbox_with_disc}
\title{Compute the intersection of a disc and its border with bounding box}
\usage{
intersect_bbox_with_disc(x, r)
}
\arguments{
\item{x}{point pattern with rectangular window}

\item{r}{radius (vector)}
}
\description{
Page 486 in IPSS08
}
