% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcf-cross-all.R
\name{ipcf_box}
\alias{ipcf_box}
\title{quick univ inhom pcf, translate border correction
assumes rectangnular window}
\usage{
ipcf_box(x, r, int, bw, adjust = 1, correction = TRUE)
}
\description{
quick univ inhom pcf, translate border correction
assumes rectangnular window
}
