% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcf-cross-all.R
\name{ipcf_cross_all_box}
\alias{ipcf_cross_all_box}
\title{Quick SOIRS Pcf In a Box}
\usage{
ipcf_cross_all_box(x, r, int, adjust = 1, correction = TRUE, verb = 0)
}
\arguments{
\item{x}{multitpye pp}

\item{r}{range vector}

\item{int}{vector of intensities, for each point}

\item{adjust}{bandwidth adjustment, bw(i,j)(x,y) = adjust * 0.15 / sqrt(lambda_j(y))}

\item{correction}{translation correction on/off}

\item{verb}{verbosity}
}
\description{
Quick SOIRS Pcf In a Box
}
