% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcf-cross-all.R
\name{ipcf_local_box}
\alias{ipcf_local_box}
\title{quick univ local inhom pcf, translate border correction}
\usage{
ipcf_local_box(x, r, int, bw, adjust = 1, correction = TRUE)
}
\description{
quick univ local inhom pcf, translate border correction
}
