% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcf-cross-all-st.R
\name{ipcf_st_cross_all_box}
\alias{ipcf_st_cross_all_box}
\title{Space-Time SOIRS Kcross-PCF}
\usage{
ipcf_st_cross_all_box(x, int, r, t, bbox, sigmas, do_correction = TRUE)
}
\arguments{
\item{x}{point pattern as an  n x (d+2) -matrix of coordinates}

\item{int}{intensity at points}

\item{r}{vector of spatial lags}

\item{t}{vector of temporal lags}

\item{bbox}{bounding box for translation correction}

\item{sigmas}{vector c(ss, st) with space and time bandwidth (sd of gaussian), respectively}

\item{do_correction}{translation correction? only sensible for cuboidal regions}
}
\value{
array of dimensions (ntype, ntype, nt, nr)
}
\description{
Space-Time SOIRS Kcross-PCF
}
\details{
Input matrix x of dimensions n x (d+2): columns 1:d taken as space dimensions, d+1 taken as the time dimension and d+2 taken as the type.

Note that the bandwidths are fixed for all pairs of types.
This is not optimal for highly imbalanced patterns.
}
