% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K-cross-all.R
\name{K_cross_all_box}
\alias{K_cross_all_box}
\title{Quick cross K, rectangle translation border correction}
\usage{
K_cross_all_box(x, r, correction = TRUE, from = NULL)
}
\description{
Quick cross K, rectangle translation border correction
}

