% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guans-kpss-test.R
\name{kpss.test}
\alias{kpss.test}
\title{Guan's KPSS test of stationarity}
\usage{
kpss.test(x, h, nx = 100, alpha = 0.05)
}
\arguments{
\item{x}{ppp}

\item{h}{variance estimation range parameter}

\item{nx}{x-step count in Rieman sums}

\item{alpha}{alpha level for test rejection. Only 0.01, 0.05, 0.1 work.}
}
\description{
Test of stationarity in a rectangle.
}

