% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf-cross-all.R
\name{markconnect_all_box}
\alias{markconnect_all_box}
\title{Quick mark connect, rectangle translation border correction}
\usage{
markconnect_all_box(x, r, ...)
}
\description{
Quick mark connect, rectangle translation border correction
}
