% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf-cross-all.R
\name{pcf_box}
\alias{pcf_box}
\title{Quick univ pcf, rectangle translation border correction}
\usage{
pcf_box(x, r, ..., int_type = "n")
}
\arguments{
\item{x}{point pattern ppp}

\item{r}{range vector}

\item{...}{passed on to rho_box}

\item{int_type}{type of intensity estimator to use.}
}
\description{
Quick univ pcf, rectangle translation border correction
}
\details{
See int_type in intensity_adapted
}

