% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf-cross-all.R
\name{pcf_cross_all_box}
\alias{pcf_cross_all_box}
\title{Quick cross 2nd pcf, rectangle translation border correction}
\usage{
pcf_cross_all_box(x, r, ...)
}
\description{
Bandwidth asymmetric: for g_ij h_ij = adjust * 0.15 / sqrt(lambda_j)
}
