% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf-cross-all.R
\name{rho_box}
\alias{rho_box}
\title{Quick univ 2nd order product density, rectangle translation border correction}
\usage{
rho_box(x, r, bw, adjust = 1, correction = TRUE, kern = 1)
}
\arguments{
\item{kern}{1=epa, 2=box}
}
\description{
Quick univ 2nd order product density, rectangle translation border correction
}
