% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf-cross-all.R
\name{rho_cross_all_box}
\alias{rho_cross_all_box}
\title{Quick cross 2nd order product density, rectangle translation border correction}
\usage{
rho_cross_all_box(x, r, adjust = 1, correction = TRUE)
}
\description{
Quick cross 2nd order product density, rectangle translation border correction
}
